Issue #2 debugging on Ubuntu platform.
The changes were originally made on a mac.  This option should now work for those of you running Ubuntu.

It could be very useful for someone with only a few probes to have directories set up by scenario, the entire scan history could be deleted and then the directory could be quick added and with just a few scans and no keypresses things would be good to go.