Issue #12 - Ready for testing
The problem with the parent window closing was due to which hbox in the
settings window the buttons were in.  The "internal action" box was
second from the bottom.  Buttons that call actions in here do not close
the window.

May need to tweak the layout to make it a little better looking but the
function should be there.